pyproject.toml
[project]
name = "aegis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.94",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "kubernetes>=35.0.0",
    "pydantic-settings>=2.14.0",
    "redis>=7.4.0",
    "strands-agents>=1.37.0",
    "uvicorn>=0.46.0",
    "websockets>=16.0",
]
